feat: add build-time generation of version information based on Git
This adds logic in the conan file to obtain version information from Git tags and pass that into CMake.
CMake, in turn, passes the information as compile-time macros to `src/version.cpp`, which creates a global with the information that can be queried at runtime.
The version information contains the commit ID, nearest tag, and whether or not there've been modifications since the tag.