feat: add support for loading textures from Targa (TGA) files
This introduces support for the R8G8B8A8 texture format to the core renderer.
In the process, an oversight was corrected where textures with 8-bit colors channels (the only supported textures at this time) were not marked as being in sRGB space. The result is that they were rendered too bright.