fix: exceptions are not always caught in the debugger
C++ exceptions that leaked to the top-level were not caught by the debugger because the `ExceptionHandler` caught them.
That logic is now disabled when a Debugger is attached, just as for native exceptions.
Additionally, a redundant try-catch block was removed which also prevented the exception from being seen by the debugger.