fix: hide private dependencies from public headers
Certain public headers were exposing headers from private dependencies (namely GLFW and Diligent). This caused build errors downstream, which cannot find those headers (since the dependencies are private).
Solve the issue by wrapping them in an Impl. For the case of Diligent, once public use had to be adopted, to replace `Diligent::NativeWindow` with a `std::any` to provide the native window to the Diligent renderer.